+++ apps/passport-server/src/services/generic-issuance/genericIssuanceService.ts
@@ -6,80 +6,8 @@ import {
 import { SerializedPCD } from "@pcd/pcd-types";
 import { Router } from "express";
 import _ from "lodash";
-import { logger } from "../util/logger";
-
-/**
- * A lemonade ticket as represented by the upcoming Lemonade API. It's
- * still t.b.d so this is my best guess for now.
- *
- * TODO:
- * - probably move to different file
- */
-export interface LemonadeTicket {
-  id: string;
-  name: string;
-  email: string;
-  eventId: string;
-  tierId: string;
-  checkedIn: boolean;
-}
-
-/**
- * A lemonade ticket tier as represented by the upcoming Lemonade API. It's
- * still t.b.d so this is my best guess for now.
- *
- * TODO:
- * - probably move to different file
- */
-export interface LemonadeTicketTier {
-  name: string;
-  id: string;
-}
-
-/**
- * A lemonade event as represented by the upcoming Lemonade API. It's
- * still t.b.d so this is my best guess for now.
- *
- * TODO:
- * - probably move to different file
- */
-export interface LemonadeEvent {
-  id: string;
-  name: string;
-  tickets: LemonadeTicket[];
-  tiers: LemonadeTicketTier[];
-  permissionedUserIds: string[];
-}
-
-/**
- * TODO:
- * - probably move to different file
- */
-export interface ILemonadeAPI {
-  loadEvents(apiKey: string): Promise<LemonadeEvent[]>;
-  checkinTicket(
-    apiKey: string,
-    eventId: string,
-    ticketId: string
-  ): Promise<void>;
-  // TODO: fill in the other methods. This is what Richard
-  // has communicated to them:
-  // - API Key scoped to an ‘account’, which has read/write permissions to the appropriate events. E.g. can read/write events that are owned/co-owned by the account.
-  // - GET product types for a given event
-  // - GET tickets for a given event, which will include
-  //     - Attendee name
-  //     - Attendee email
-  //     - Product type (7-day, early bird, GA, etc)
-  //     - Checked-in status
-  // - POST Check-in (and potentially check-out)
-}
-
-export interface LemonadeUser {
-  id: string;
-  email: string;
-  apiKey: string;
-  name: string;
-}
+import { ILemonadeAPI } from "../../apis/lemonade/lemonadeAPI";
+import { logger } from "../../util/logger";
 
 /**
  * TODO:
